+++ mne/tests/test_epochs.py
@@ -14,7 +14,8 @@ import numpy as np
 import copy as cp
 import warnings
 
-from mne import fiff, Epochs, read_events, pick_events, read_epochs
+from mne import (fiff, Epochs, read_events, pick_events, read_epochs,
+                 equalize_channels)
 from mne.epochs import bootstrap, equalize_epoch_counts, combine_event_ids
 from mne.utils import _TempDir, requires_pandas, requires_nitime
 from mne.fiff import read_evoked
@@ -22,6 +23,7 @@ from mne.fiff.channels import ContainsMixin
 from mne.fiff.proj import _has_eeg_average_ref_proj
 from mne.event import merge_events
 from mne.externals.six.moves import zip
+from mne import equalize_channels
 
 warnings.simplefilter('always')  # enable b/c these tests throw warnings
 
@@ -186,3 +188,17 @@ def test_drop_channels_mixin():
     epochs.drop_channels(drop_ch)
     assert_equal(ch_names, epochs.ch_names)
     assert_equal(len(ch_names), epochs.get_data().shape[1])
+
+def test_equalize_channels():
+    """Test equalization of channels
+    """
+    epochs1 = Epochs(raw, events, event_id, tmin, tmax, picks=picks,
+                    baseline=(None, 0))
+    epochs2 = epochs1.copy()
+    ch_names = epochs1.ch_names[2:]
+    epochs1.drop_channels(epochs1.ch_names[:1])
+    epochs2.drop_channels(epochs2.ch_names[1:2])
+    my_comparison = [epochs1, epochs2]
+    equalize_channels(my_comparison)
+    for e in my_comparison:
+        assert_equal(ch_names, e.ch_names)