+++ doc/source/whats_new.rst
@@ -28,6 +28,9 @@ Changelog
    - Add 3D MEG/EEG field plotting function and evoked method by `Denis Engemann`_ and  `Alex Gramfort`_
 
    - Add consistent channel-dropping methods to Raw, Epochs and Evoked by `Denis Engemann`_ and  `Alex Gramfort`_
+   
+   - Add `equalize_channnels` function to set common channels for list of
+   	 Raw, Epochs, or Evoked objects by `Denis Engemann`_
 
 API
 ~~~